requirement.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
Django==5.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
ffmpeg-python==0.2.0
fonttools==4.56.0
ftfy==6.3.1
future==1.0.0
gallery_dl==1.28.5
humanize==4.12.1
idna==3.10
ipython==8.32.0
jedi==0.19.2
kiwisolver==1.4.8
library==3.0.59
LibraryFile @ file:///C:/Users/ksara/OneDrive/Desktop/PythonClassEve/MyLibraryOne
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mutagen==1.47.0
mylib==0.0.1
# Editable install with no version control (MyLibrary==1.0)
-e c:\users\ksara\onedrive\desktop\pythonclasseve\my_library
natsort==8.4.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
puremagic==1.28
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
pyreadline3==3.5.4
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-mpv-jsonipc==1.2.0
pythonnet==3.0.5
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
screeninfo==0.8.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.1
stack-data==0.6.3
tabulate==0.9.0
tinytag==2.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
VedAstro==1.23.18
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0
yt-dlp==2025.2.19
